리팩토링

+++ src/main/kotlin/com/njhyuk/swagger/rag/domain/model/ApiDocChunk.kt
@@ -4,5 +4,5 @@ data class ApiDocChunk(
     val path: String,
     val method: String,
     val text: String,
-    val embedding: List<Float>? = null
+    val embedding: List<Double>? = null
 )
